Update plugin name to use underscores
The opa --set flag does not support periods in keys which means we
cannot define the plugin configuration on the command line. This
commit updates the plugin name to envoy_ext_authz_grpc instead of
envoy.ext_authz.grpc so that the plugin can be configured on the
command line.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>