config: Add a new parameter that specifies the policy decision path
These changes deprecate the "query" parameter and add a new "path" parameter to specify the policy decision path. Earlier the query name was used to generate the policy decision path in decision logs. With changes to the decision logger api to no longer modify the decision path, the plugin now sets the appropriate value for the path and query in the decison log.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>