internal: Updates to generated decision logs and non-OPA error wrapping
This commit updates the server handler to create a decision ID and
a Metrics object on start-up that can be later included in the decision log.
This ensures that the decision log will have an ID and the server handler time
even in error scenarios.

This change also wraps non-OPA errors to ensure they are properly
serialized.

Fixes open-policy-agent/opa#2591, Fixes open-policy-agent/opa#2592,
Fixes open-policy-agent/opa#2593, Fixes open-policy-agent/opa#2594

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>