CI: Add GHA to create GH releases and push binaries
This commit adds a GitHub Action workflow that builds and publishes
release binaries and creates/updates a release on GitHub. The binaries are
created for the tagged version. This workflow is trigerred after a tag has been pushed.

The scripts utilized in this commit are modified versions of those used for a similar workflow
in the core OPA repo.

Fixes open-policy-agent/opa#2597

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>