Makefile: Use the plugin version to build releases
Earlier we used only the OPA version to create a GH release and
build the release binaries. We did not create a new GH release when
changes we made to the plugin itself.

This change updates the release target to use the plugin version
instead of only the OPA version to checkout and build binaries.
This will allow us to create releases for the plugin revisions.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>