internal: Handle errors from OPA eval cancellation
When the context is cancelled or deadline is reached
durign query evaluation, the error message returned is
"caller cancelled query execution". This message is
not very informative from the pov of the plugin user.
This change updates the plugin to handle eval cancellation
errors and returns a more helpful error message.

This change also checks if the context is cancelled or
deadline is reached even before policy eval has started
which may happen in scenarios where the server is overloaded and
has pending requests in its queue and returns an appropriate
error message in such cases.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>