Support Envoy's v3 xDS API, envoy.service.auth.v3.Authorization
This change makes opa-envoy-plugin expose two gRPC services:
envoy.service.auth.v2.Authorization and envoy.service.auth.v3.Authorization.

Since switching to v3 requires user action already, we've taken this chance
to migrate the representation of the v3 CheckRequest to the specified JSON
mapping. See the README.md for details.

That aside, if you're using the v2 API, no changes to your policies are
required, and you can keep using the v2 API for some time. It's TBD when we
are going to drop support for it.

This change also brings another input value that lets you determine the
request version (v2/v3) and its encoding (protojson for the official mapping,
"encoding/json" for the one used with v2).

With respect to the software we're integrating with here, note

- Envoy 1.13 is the first version to support the v3 API.
- Istio 1.7.0 is the first version that lets you configure the v3 filters
  required to use the v3 API with Envoy.
  Support for the 1.6 series of Istio has been ended in Nov 21 2020.
- That first version of Envoy released in 2021, presumably 1.18, will drop
  support for the v2 API.

The `examples/istio/quick_start.yaml` file has been adapted:

1. it's using the v3 API
2. its hardcoded TLS certs (generated with `build/gen-tls-certs.sh`) work
   with a service that's built with Go 1.15 (which refuses CN in favour of
   SNI records)

Signed-off-by: Stephan Renatus <stephan@styra.com>