workflow: use docker-env for building image
Before, the `image` would invoke `build-linux`, and that wouldn't guarantee the
version of golang used for building.

Now, we're following the schema used by OPA to build the binary into a release dir,
and have the image build pick it up.

Also, the built image now includes everything needed to use wasm:

    $ docker run  -it docker.io/openpolicyagent/opa:0.25.1-envoy-2 version
    Version: 0.25.1-envoy-2
    Build Commit: 20832273
    Build Timestamp: 2020-12-08T13:34:55Z
    Build Hostname: ce578c1a1309
    Go Version: go1.15.2
    WebAssembly: available
    $ docker run --entrypoint busybox -it docker.io/openpolicyagent/opa:0.25.1-envoy-2 sh
    / # ls /usr/lib/opa
    libwasmer.so
    / #

We do this by building off of the openpolicyagent/opa package, which has the
library in place.

Fixes https://github.com/open-policy-agent/opa/issues/2942.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>