Fix unmarshal json when Content-Type is application/json
It should not be able to parse yaml when Content-Type is application/json.
However `util.Unmarshal` is a utility that can parse yaml or json,
so if you put yaml in the body when Content-Type is application/json, it will be parsed.
Therefore, change the calling function to a `util.UnmarshalJSON`, and make it possible to parse only json.

In addition, `util.Unmarshal` was not possible to unmarshal a json string containing `\/` (which is an escaped `/`).
Json is supposed to consider slashes (`/`) as escapable characters (not required).
It could not parse a json body containing string of `\/`.
Because of the above, the opa-envoy-plugin eventually returned 403
whenever the Content-Type was application/json and the json contained `\/` in the string.
You can also avoid this problem by using the `util.UnmarshalJSON`.

Signed-off-by: Daisuke Takada <tkdtkd0022@gmail.com>