build: update tag script from github action
This allows us to use dependabot for updating OPA.

For any merge, we'll check if the version of OPA has changed,
and if so, push a commit fixing the version tags in the example
yamls.

It's done in a rather blunt manner: we replace whatever version
is in the yamls with whatever version we have for opa in go.mod,
and if the result is different from what we had before, we'll
push it to main.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>