Allow option to bypass request body parsing
This change allows users to bypass request body parsing
by exposing a new config option on the plugin to control
this. Sometimes admins for ex. may not have control over the
request body in the input request and may still want
OPA to evaluate the request instead of the plugin
returning an error.

Fixes: #open-policy-agent/opa/issues/5081

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>